+++ Bencodex.Tests/Types/NullTest.cs
@@ -1,0 +1,14 @@
+using Bencodex.Types;
+using Xunit;
+
+namespace Bencodex.Tests.Types
+{
+    public class NullTest
+    {
+        [Fact]
+        public void Value()
+        {
+            Assert.IsType<Null>(Null.Value);
+        }
+    }
+}