+++ Bencodex/Types/Null.cs
@@ -12,6 +12,15 @@ namespace Bencodex.Types
         IComparable<Null>,
         IComparable
     {
+        /// <summary>
+        /// Represents a <see cref="Null"/> instance.  Recommends to prefer this over using
+        /// the default constructor or a <c>default</c> keyword.  This field is read-only.
+        /// </summary>
+        public static readonly Null Value =
+#pragma warning disable SA1129
+            new Null();
+#pragma warning restore SA1129
+
         [Pure]
         public string Inspection => $"null";